Fixed preference base lists not intialising before rendering relevant pages (should still generate recipes if not opening any preference pages)
