+++ frontend/src/components/Dropdown.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import { useEffect, useState, useContext } from 'react';
-import DropdownPane from "./DropdownPane"
 import { PreferenceContext } from '../context/PreferenceContext';
+import { baseCuisineList } from '../utils/preferenceBaseLists';
+import DropdownPane from "./DropdownPane"
 
 function Dropdown({}) {
   const { prepTime, setPrepTime } = useContext(PreferenceContext);
@@ -36,38 +37,9 @@ function Dropdown({}) {
   }
 
   //Cuisine List Handlers
-  const baseCuisineList = [
-    {id: 0, name: "African", value: "african", state: false},
-    {id: 1, name: "Asian", value: "asian", state: false},
-    {id: 2, name: "American", value: "american", state:false},
-    {id: 3, name: "British", value: "british", state: false},
-    {id: 4, name: "Cajun", value: "cajun", state: false},
-    {id: 5, name: "Caribbean", value: "caribbean", state: false},
-    {id: 6, name: "Chinese", value: "chinese", state: false},
-    {id: 7, name: "Eastern European", value: "easterneuropean", state: false},
-    {id: 8, name: "European", value: "european", state: false},
-    {id: 9, name: "French", value: "french", state: false},
-    {id: 10, name: "German", value: "german", state: false},
-    {id: 11, name: "Greek", value: "greek", state: false},
-    {id: 12, name: "Indian", value: "indian", state: false},
-    {id: 13, name: "Irish", value: "irish", state: false},
-    {id: 14, name: "Italian", value: "italian", state: false},
-    {id: 15, name: "Japanese", value: "japanese", state: false},
-    {id: 16, name: "Jewish", value: "jewish", state: false},
-    {id: 17, name: "Korean", value: "korean", state: false},
-    {id: 18, name: "Latin American", value: "latinamerican", state: false},
-    {id: 19, name: "Mediterranean", value: "mediterranean", state: false},
-    {id: 20, name: "Mexican", value: "mexican", state: false},
-    {id: 21, name: "Middle Eastern", value: "middleeastern", state: false},
-    {id: 22, name: "Nordic", value: "nordic", state: false},
-    {id: 23, name: "Southern", value: "southern", state: false},
-    {id: 24, name: "Spanish", value: "spanish", state: false},
-    {id: 25, name: "Thai", value: "thai", state: false},
-    {id: 26, name: "Vietnamese", value: "vietnamese", state: false}
-  ]
 
   const [cuisineList, setCuisineList] = useState(() => {
-    if (cuisine === undefined) return baseCuisineList;
+    // if (cuisine === undefined) return baseCuisineList;
     return cuisine;
   });
